move hubblestack.splunklogging to hubblestack.log.splunk (and other related fixes)
