+++ hubblestack/log/__init__.py
@@ -1,0 +1,9 @@
+# coding: utf-8
+
+from hubblestack.log.setup import (
+    LOG_LEVELS,
+    emit_to_splunk,
+    setup_splunk_logger,
+    setup_console_logger,
+    setup_file_logger
+)
